Extract local and remote operations into respective source files.
